requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==3.2.15
django_debug_toolbar==3.8.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
py==1.11.0
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytils==0.4.1
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
